requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alibabacloud-ocr-api20210707==3.1.2
alibabacloud-tea==0.4.0
alibabacloud_credentials==0.3.6
alibabacloud_darabonba_stream==0.0.2
alibabacloud_endpoint_util==0.0.3
alibabacloud_gateway_spi==0.0.2
alibabacloud_openapi_util==0.2.2
alibabacloud_tea_openapi==0.3.12
alibabacloud_tea_util==0.3.13
alibabacloud_tea_xml==0.0.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
azure-ai-documentintelligence==1.0.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
dashscope==1.22.1
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
gradio_client==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
imageio==2.37.0
isodate==0.7.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msrest==0.7.1
multidict==6.1.0
narwhals==1.24.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.2
oauthlib==3.2.2
openai==1.59.9
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pikepdf==9.5.1
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
PyPDF2==3.0.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.3
safehttpx==0.1.6
scikit-image==0.25.0
scipy==1.15.1
semantic-version==2.10.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.45.3
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
tifffile==2025.1.10
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
